+++ src/Tabs.tsx
@@ -8,12 +8,14 @@ interface TabProps {
     cy?: string
     test?: string
   }
-  key: string
+  key?: string
   value: string
   label: string
+  disabled?: boolean
   className?: {
     root?: string
     label?: string
+    disabled?: string
   }
 }
 
@@ -26,10 +28,19 @@ interface TabProps {
  * @param key - The key of the tab.
  * @param value - The value of the tab. This is required for the internal and external state.
  * @param label - The label of the tab.
+ * @param disabled - The optional disabled property allows you to disable the tab.
  * @param className - The optional className object allows you to override the default styling.
  * @returns Tab trigger component
  */
-export function Tab({ id, data, key, value, label, className }: TabProps) {
+export function Tab({
+  id,
+  data,
+  key,
+  value,
+  label,
+  disabled,
+  className,
+}: TabProps) {
   return (
     <TabsPrimitive.Trigger
       id={id}
@@ -43,6 +54,7 @@ export function Tab({ id, data, key, value, label, className }: TabProps) {
         'focus:rdx-state-active:border-b-red focus:z-10 focus:outline-none focus-visible:ring focus-visible:ring-purple-500 focus-visible:ring-opacity-75',
         className?.root
       )}
+      disabled={disabled}
     >
       <span
         className={twMerge(